Add react-bmi-calculator as a test (#16)
We also have this as a separate repo of course, but having it as part of
report-diffs-action ensures that the tests run using the version of the
action on the pull request, rather than the last published version. I've
put in a heavily cut down version of the react-bmi-calculator, so the
number and size of the files is small.

The github action isn't really designed to work with
useAssetsSnapshottedInBaseSimulation (it doesn't make so much sense when
comparing against the base commit of a PR), so in
https://github.com/alwaysmeticulous/report-diffs-action/pull/14 I'm not
exposing that argument. Hence the need to change how we do the tests,
and run them against a real URL.