Report number of screens that differ, rather than number of sessions (#26)
This PR blocks on
https://github.com/alwaysmeticulous/meticulous-sdk/pull/179.

The intent is that what the user sees in GitHub should line up with what
they see when they click the link: if they see 6 screenshots when they
click the link, then they should see the number 6 in GitHub. And if no
screenshot diffs show up when they click the link, then they shouldn't
see it as failed in GitHub.

In the medium/long term we'll want to better surface other kinds of
errors, rather than just hide them. Designs for this in Figma.