Fix cloud compute action head SHA
This fixes a bug where the head SHA provided as input was ignored in the
cloud compute action.

Additionally, this extends the action to deal with cases where the head of
the PR commit is being tested, rather than the temporary merge commit.
Previously we used the temp merge commit as a head commit sha regardless
of what was being checked out.