Use BE endpoint for calculating base in cloud-compute action (#454)
This PR updates the base commit calculation to use a BE endpoint instead
of doing it within the action. This aligns us

The logic within the "old" action is kept as is as to maintain ability
to use without installing the GH app.

This also drops the 'head sha' parameter from the old action. It's not
currently used and it's usage is incorrect.