Always trigger a test run for a missing base to the base code change commit (#493)
This is the default for non-monorepo enabled projects. If we have
fallback to an earlier commit enabled we might try to kick off a
workflow for an older commit which we can't do so instead of we use the
commit from the code change event which is guaranteed to be same or
newer.